feat: character card additional info
A character card with additional information has been created. Also, when transitioning to the page with additional information, the transfer of received data for a specific character is performed, which reduces the number of requests. However, at the same time, requests for additional information begin to be executed upon transition. Icons have also been added to display the model of the ship and the number of passengers.